requirements.txt
APScheduler
PyYAML
click
watchdog
